backend/requirements/base.txt
Django==4.2.*
djangorestframework==3.14.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
channels==4.0.*
channels-redis==4.1.*
daphne==4.0.*
psycopg2-binary==2.9.*
redis==5.0.*
python-dotenv==1.0.*
requests==2.31.*
razorpay==1.4.*
cloudinary==1.36.*
django-cloudinary-storage==0.3.*
whitenoise==6.6.*
dj-database-url==2.1.*
gunicorn==21.2.*
